Delete routes directory and directly connect components to routes
