+++ src/App.js
@@ -1,9 +1,8 @@
 import { Routes, Route } from 'react-router-dom';
 import './App.css';
-import Rockets from './routes/Rockets'
-import Missions from './routes/Missions'
-import Profile from './routes/MyProfile';
-import NotMatch from './routes/NotMatch';
+import Rockets from './components/Rockets';
+import Missions from './components/Missions';
+import Profile from './components/myProfile';
 import Layout from './components/Layout';
 
 
@@ -14,7 +13,6 @@ function App() {
       <Route path="/" element={<Rockets />} />
       <Route path="missions" element={<Missions />} />
       <Route path="profile" element={<Profile />} />
-      <Route path="*" element={<NotMatch />} />
       </Route>
     </Routes>
   );
