+++ src/components/Missions.js
@@ -1,0 +1,7 @@
+const Missions = () => {
+    return(
+        <></>
+    )
+}
+
+export default Missions
