+++ src/components/Rockets.js
@@ -1,0 +1,7 @@
+const Rockets = () => {
+    return(
+        <></>
+    )
+}
+
+export default Rockets
