+++ src/components/myProfile.js
@@ -1,0 +1,7 @@
+const Profile = () => {
+    return(
+        <></>
+    )
+}
+
+export default Profile
