--- src/routes/Missions.js
@@ -1,7 +1,0 @@
-const Missions = () => {
-    return(
-        <></>
-    )
-}
-
-export default Missions
