--- src/routes/MyProfile.js
@@ -1,7 +1,0 @@
-const Profile = () => {
-    return(
-        <div>This is my profile.</div>
-    )
-}
-
-export default Profile
