--- src/routes/NotMatch.js
@@ -1,8 +1,0 @@
-const NotMatch = () => {
-    return (
-      <div>
-        <p>Not found!</p>
-      </div>
-    );
-  };
-  export default NotMatch;
