--- src/routes/Rockets.js
@@ -1,7 +1,0 @@
-const Rockets = () => {
-    return(
-        <></>
-    )
-}
-
-export default Rockets
